Round trip the calculation output double to an integer before converting to a string to eliminate the extraneous decimal

+++ lib/main.dart
@@ -72,7 +72,7 @@ class _CalculatorPageState extends State<CalculatorPage> {
             }
           }
 
-          displayText = result.toString();
+          displayText = result.toInt().toString();
           firstOperand = "";
           operator = "";
           secondOperand = "";
